+++ source/js/components/countryListHeader.ts
@@ -1,0 +1,8 @@
+import {c, msgid} from '../tools/translate';
+
+export const countryListHeader = (count: number) =>
+	/* translator: Header of a list of countries */ c('Label').plural(
+		count,
+		msgid`Country`,
+		`Countries`,
+	) + (count > 1 ? ' (' + count + ')' : '');